refactor(app): decompose the 3.4k-line BossApp.kt into focused modules
BossApp.kt had grown to 3,367 lines doing a dozen jobs at once. It is
now an 83-line composition root that only wires layers together; every
concern moved to its own file with an explicit seam:

ai.rever.boss.app (new package):
- BossAppState: per-window component graph + all transient UI state
  (dialog flags, wizard progress, startup coordination), built by
  rememberBossAppState; handler-readiness CAS moved behind
  markHandlersReady()
- BossAppStartupEffects: lifecycle wiring — registries, pending
  tab/project consumption, DefaultPlugin lifecycle, workspace restore +
  handler readiness, install wizard, updater/CLI bootstraps,
  performance providers, workspace auto-save
- BossAppEventBusEffects: the window-filtered event-bus listeners
  (file/terminal/runner/git/link/run/workspace/panel/dashboard/URL)
- BossAppMenuActionEffects: MenuActionsHandler listeners + derived
  menu state (split enabled, panel count)
- BossAppScaffold + BossAppCompositionLocals: window chrome, overlays,
  and the CompositionLocal surface plugins read
- BossAppDialogs: all 14 dialogs, driven by BossAppState flags
- FocusModeReveal: the 4-edge hover-reveal state machine + strips
- TabFactories / TerminalLinkOpener: tab creation + terminal-link
  opening helpers

ai.rever.boss (kept for compatibility):
- PlatformHooks.kt: the expect declarations (actuals unchanged)
- TabDropHandler.kt: handleTabDropResult, same package so its unit
  test and internal visibility are untouched

Pure code movement — no behavior changes, public BossApp signature
unchanged. Only deletion: the `saveMessage` snackbar state, which was
never written anywhere (dead since extraction).
